temu-web-app/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pandas>=1.5.0
openpyxl>=3.0.10
chardet>=4.0.0
colorama>=0.4.4
python-dotenv>=1.0.0
pyunpack>=0.3.0
patool>=1.12
py7zr>=0.20.5
unrar>=0.4
